pdf_api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
pdfplumber==0.10.2
PyMuPDF==1.23.8
Pillow==10.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
numpy==1.26.0
